Fixes new ROL interface to reset nonlinear solver between sweeps
A behavioral difference was observed between the old and new ROL
implementations.  The convergence of the simple diode example had
a different history, when it really shouldn't.  This was tracked
down to a need to reset the nonlinear solver object between the
DC sweeps.  For NOX, there is internally stored information that
needs to be cleared between these sweep calls.  Without clearing
the nonlinear solver, the next sweep will be different even if the
nonlinear problem is the same.

I've also changed the pattern in the ROL analysis to register the
active analysis object with the analysis manager, similar to HB.
ROL is more or less a strategy object that uses individual analysis
objects that may be active at different times to evaluate the
objective or constraints for the optimization algorithm.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/370