Fix the Hspice-style limit operator
Issue #377
"The limit random operator (Hspice 2-argument version) is
not correct in the expression library"

This came up while investigating another issue.  The
limit(nominal,abs_variation) operator has never been correct in Xyce
until now.  It was incorrect for both the sampling and non-sampling
use cases.

In the new expression library, I had implemented all the Hspice random
operators correctly except this one.  For this one I mimiced what
had been set up in the old library, and then didn't do much with it
after that.  Unfortunately, the old library was
actually doing the wrong thing, even  for the non-sampling case.
And of course, the old library was never connected to sampling, to
support the sampling use case.

Here is some more detail.

Xyce supports two different versions of limit.  One is the Pspice
version, which takes 3 arguments.  The other is the Hspice version,
which takes 2 arguments.  (technically in Hspice, limit also has an
optional 3rd argument, but I'm ignoring that for now).

This commit concerns the 2-argument Hspice version.  That version
supports corner analysis via sampling.  The two arguments are
"nominal" and "abs_variation" specified as:

    limit(nominal,abs_variation)

If running WITH sampling, then this operator returns either
(nominal+abs_variation) or (nominal-abs_variation), with equal
probability.  Prior to this commit, Xyce did not do this,
b/c this operator wasn't properly connected to the various
sampling methods.  Instead it behaved the same as the rand()
operator.  That has been fixed in this commit.

If running WITHOUT sampling, this operator should return the
nominal value.  However, until today it was returning
(nominal+abs_variation).  That was incorrect, and the
expression library has been updated to return nominal for the
non-sampling case in this commit.