Reformat LaTeX source and reorder fixed defects
In reviewing an issue, I wanted to clean up the formatting of the
latex source code for that issue's entry in the fixed defects
table (making it suitable for editing in an 80-column window, removing
whitespace errors).  In doing so I wound up finding other entries to
clean up, and did so.

I have also reordered the table so that the issues are presented in
order of increasing issue number.  This is completely pointless, and
something we have rarely bothered to keep consistent, but in some
releases we have made sure the list is sorted.

In sorting the list, I found one issue that had been documented twice
in exactly the same way.  Hopefully, having the list sorted will make
it easier to spot that one has already put an entry in before putting
it in again.