Fix HB output processing to work with APFT and box truncation
The box truncation uses unordered list of harmonics. This is because
the frequency mapping based HB needs the mapped frequencies be ordered.
The real frequencies are not ordered. In order to correctly output
the HB results for APFT with box truncation method, the output processing
needs to be changed

Xyce#231 Xyce#357