Removes several vector initializations
There are several uses of putScalar in the AC code followed by an
update on the same vector.  These two methods can be consolidated
if the scalar for the vector performing the update is zero.  This
reduces the number of times the memory is touched.  While it is not
the reason for the expensive dotProduct in the AC sensitivity code,
it is still a best practice.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/364