Added Xyce::Simulator::setCircuitParameter()
Added a method to allow the controlling program (such as Simulink)
to set a simulation parameter by name.  This calls the device
manager through the Xyce::Simulator class but that's hidden
from the caller.  The caller just get's a boolean false if
the parameter doesn't exist and a true if it was set.  In the
test functions in XyceSimulatorUnitTests.C there is a test,
SetDeviceParamTestNetlist2, that trys to set existing
and nonexisting parameters and tests that it gets the
right return result.  Also, parameters are modified during
a transient simulation and that works.  The test doesn't verify
that Xyce's calculations reflect the new parameter values, but
one can do that by looking at the PRN output file.  Doing that
as well is probably a good idea.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/104