Creates a new method applyOmega_dJdp to address performance issue
This commit includes a new method that computes a matrix-vector product
using a scaled version of a block matrix (or its transpose) without
assembling it.  Previously, the block matrix was scaled and reassembled
to perform the matrix-vector product and then the block matrix was
unscaled and reassembled.  This was somewhat expensive for a small
circuit in serial, but really caused performance issues for the same
circuit in parallel.  With this modification, the serial and parallel
runtimes are about the same.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/364