Fix for Issue #306, to speed up expression-based sampling
Issue #306
"Using expression random operators with sampling is too slow for large
numbers of parameters"

This commit contains a lot of changes.  Most importantly, for
expression-based sampling, Xyce no longer does a lot of redundant
work.  It accomplishes this by carefully tracking exactly which device
parameters depend on exactly which .param and/or .global_params, and
ONLY doing the bare minimum updates.  Tracking these dependencies
required updates to the various setDependentParameter functions in the
device entity.  There is also more information in the struct
"UserDefinedParams", which is found in the N_DEV_SolverState.h header.

As part of this work, I have also created a new function
called Device::setParameterRandomExpressionTerms2.

It is similar to the original ::setParameterRandomExpressionTerms,
except that it handles *all* the sampling parameters in one function
call, rather than one by one.  This original function is now obsolete.

In order to make this work, and also to plan for some of my future work,
I had to add extra information to the SweepParam class. (which is used
by both sweeps and sampling).  One benefit of the extra information is
that I don't need to do as many searches.

The new function Device::setParameterRandomExpressionTerms2  only
re-evaluates expressions and resets device parameter values (via the
updateGlobalAndDependentParameters function call) at the end of the
global parameter loop.  By doing it this way, I avoid redundant
expression evaluations, which for the test circuit are very expensive.

The new code also filters out .param/.global_param that do not have any
device model or instance dependencies.  Before this work I had
mistakenly assumed that this sort of filtering happening during parsing,
but I was wrong about that.  I had to be extra careful do to this
filtering properly in parallel, and check for dependencies on all
processors, not just the local processor.

Another aspect that was important for parallel simulation was the order
of global parameters.  Global parameters are not necessarily in the same
order on each processor.  For previous implementations of sampling
and step, I had put in enough sorting routines to make them consistent.
However, the new method, needed more sorting to work correctly in parallel.

With all of these changes, the original test problem runs a LOT faster.