Fix pause breakpoint at time = 0 from DC OP
The Xyce::Simulator::SimulateUntil() method would fail if the
calling program asked for Xyce to start-up and simulate until time=0.
The issue was that while a pause breakpoint was correctly set at
time=0, the doRun() method of N_ANP_Transient.C would try to
take one time-step after the DC OP before checking for a pause
breakpoint.  The logic in Transient::doLoopProcess() assumed
that Xyce was not at a breakpoint before taking a time step, so it
trys to take one and then checks for breakpoints.  The solution was
to check for pause breakpionts after the DC OP and before starting
doLoopProcess() which is all about taking transient steps correctly.

I also updated XyceSimulatorUnitTests.C so that test
MultiTimeStepTestNetlist1 specificaly requests that
Xyce::Simulator::SimulateUntil() goes to time 0.0.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/104