Remove N_CIR_MixedSignalSimulator usage
I made the Simulink sfunction interface use a xyce object
created from the MixedSignalSimulator class early on in this
work. I later learned that the MixedSignalSimulator class has issues
that make it unsuitable for this work.  With the last commit on
making Xyce::Simulator::SimulateUntil() work for getting just the
DC op it is possible to drop the MixedSignalSimulator derived
class and just use Xyce::Simulator.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/104