Update AC SENS parameter options in RG and UG
It is now possible to use both acobjfunc and objvars on the same .SENS
line to specify the sensitivity objective functions for AC (previously,
the .sens line could only contain one or the other). This restriction
was removed as part of issue #270 and implemented in commit
90261cbe5848ac690680f978b99191ef5c6a0cd0. This commit documents the new
behavior in the Users Guide and Reference Guide.

This commit also removes an extra pair of {} from the standard output
option for AC sensitivity. The column headers for sensitivity output
when through a few revisions for issue #270. In the final version, the
curly braces were incorporated into the function name and I forgot to
remove the ones from the standard out print function.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/270