Apply new UQ param handling to non-sampling UQ
This is related to issue #306, which was all about speeding up the
parameter handling for .SAMPLING.

For that work, among other things I created a new version of the
funcntion

DeviceMgr::setParamRandomExpressionTerms

and I called it:

DeviceMgr::setParamRandomExpressionTerms2

The old function was obsolete, so I removed it from
Xyce, as well as the call trees thru the loader leading to it.

In addition to removing the old function, this commit also applies the
new function to the non-.SAMPLING UQ methods.  Specifically
.EMBEDDEDSAMPLING and .PCE.    As a result,
the new parameter handling will benefit them as well,
and I can safely remove the obsolete code.