Update user guide src-stepping section for Isrcs
Issue #363
"Extend source stepping in Xyce to apply scalar to ISRC devices as well
as Vsrc devices"

The user guide section on source stepping was worded to be
Vsrc-specific.  This commit changes it to include mention of current
sources as well.

This commit also adds documentation of issue #363 to the release notes,
in the "robustness and stability enhnacements" section.