Allow Simulate to access voltage nodes on output
The access function in Xyce::Simulator that allows an external
program to get the value of solution variable like a voltage node was
not correctly making the operator needed.  This fixes the operator
creation and updates the unit test that was failing to catch this.
Also the simulation interface code has been updated to so it correctly
requests solution variables from Xyce's output.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/104