Add better support for xyce-charon coupled builds
When Trilinos is built with TriBITS (as it is for Charon), the usual
Trilinos_ prefix on a lot of the cmake variables are changed (the
new prefix reflects whatever the overall TriBITS project is). This
commit adds logic to Xyce's the top level CMakeLists file to check
for a new parameter that indicates if it's doing a TriBITS project
build (specifically tcad-charon). If it is a TriBITS build, the
Trilinos properties are now explicitly set based on the corresponding
TriBITS properties. This way, the Trilinos references in the other
cmake files don't need to be updated.

I also encountered some errors while building Charon coupled to Xyce,
apparently due to Xyce headers that are referenced in Charon but are
not part of the Xyce install. I have added these headers to the Xyce
install (in the xyce-for-charon build only), but we may need to follow
up and see why these headers are needed.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/266