Improve naming in CMake to match convention
When I recommended the approach taken by Josh, I gave him the variable
name, `TB_prefix`. While that works, it doesn't match the convention
used in this project, which is consistent with the convention used by
the CMake project. Therefore, this commit changes the cache variable
name to `TriBITS_prefix`. I also put the "tcad-charon" header file list
in alphabetical order and added a comment.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/266