Fix for 2-level debug output.
Xyce#379

When running in debug mode, the "AnalysisManager::getStepNumber"
function is used to provide indices to matrix and vector files.
These are output from the nonlinear solver, and this is only enabled
when the netlist contains ".option nonlin debuglevel=2" or something
like that.

Prior to this commit, this capability was slightly broken when invoked
from the inner solve of a 2-level solve.  It was broken in that the step
index, which was used in the various filenames, was never updated
properly. As a result most files got overwritten.

It took a little digging but I figured out why this was. The reason had
to do with a refactor that happened to the analysis package a number of
years ago. That refactor was looking ahead toward managing multiple
analyses in a single netlist.

To support this, the analysis manager stores a vector of one or
more analysis objects, with one of those being labeled as
the "primary" analysis object.  (usually the "back" object in the
vector).

However, when running 2-level, the inner solve was incorrectly
allocating an additional analysis object.  This wasn't necessary, as it
was redundant with the already-allocated primary analysis object. Then
after allocating this redundant object, it would call
"setPrimaryAnalysis" function so it would be the preeminent analysis
object.  But, when this function is called, it doesn't get put into the
correct location in the analysis stack.  Some functions (such as
getStepNumber, used by the debug outputs) were sensitive to this
mistake, and some where not.

As a result, there were (sort of) 2 primary analysis objects
in the inner solver, and some functions called one and some called the
other. This got overlooked b/c it didn't break any 2-level regression
tests. But it caused the debug output to be mangled.

This commit fixes the code that allocated another analysis object, to
simply grab the existing primary analysis object.  To do this I had to
add a protected function to the analysis manager and change a handful of
lines in the SecondLevelManager.

If not using 2-level, these changes won't change anything. So this is a
relatively low-risk change.