Further improve the CMake against TriBITS
It was pointed out that the mixed-mode simulation capability requires a
different probe of the Trilinos setup. I think this is the solution.
The following are taken from comments in the CMake files, which are
worth replicating here as a record of what we had to do and why.

A TriBITS project that performs a coupled build with Trilinos, like
Charon, changes the names of all the Trilinos CMake variables.
Therefore, we have to jump through a bunch of hoops to take that use
case into account. So we set a cache variable to handle that name
"mangling," called `TriBITS_prefix`. For example, set
`TriBITS_prefix="tcad-charon"` to build against Charon. A section early
in the CMake process ensures the variables all start with "Trilinos" so
nothing special has to be done later in the process.

The version check for Trilinos is behind an if() statement due to the
TriBITS name "mangling" issue. From what I tell, there is no way to
probe the version of Trilinos when it is built as part of a different
Tribits project.

Charon can use a specific part of Xyce as a TPL. However, that use case
results in a build of Xyce that is unusable for any other purpose.
Therefore, we need to make a special build of Xyce that skips some
checks. To do that, `Xyce_AS_SPECIAL_CHARON_TPL=TRUE` has to be set at
configure time. This is _not_ a cache variable, because I don't want
anyone trying to use it who doesn't know exactly what they are doing.

When Charon uses that specific part Xyce as a TPL, it links to code deep
in Xyce. Therefore, it requires a lot more headers to be installed for
proper linking. We're installing those headers for all Charon builds,
because it is not worth dealing with yet another special use case.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/266