Wire L-UTSOI into Xyce
The 18 month CMC embargo on release of L_UTSOI expired this month, and
we are migrating the code from Xyce_NonFree to Xyce so that it can be
present in open source builds of the code.

A previous commit moved the Verilog-A source code from one repo to
another.  This commit simply changes the build system so that the
model is built from the moved copy.

Xyce#378