Implement 2-level DAE vector outputs
Xyce/Xyce#379

This issue is about extracting DAE information (mainly f and q vectors)
during a 2-level solve, from one or more of the "inner" solvers.

The initial work has been to try and leverage the debug output functions
in the nonlinear solver.  But this is a blunt instrument, and produces a
lot more output than desired.  Also, it requires the user to have
compiled a debug version of Xyce.  And finally, the files often
overwrite each other (related to the excess output).

To address these issues, this commit adds a new set of output
functions, as well as netlist commands to invoke them.  The new
netlist commands are:

options twolevel
+ output_dae_vectors=true
+ output_dae_matrices=true

which would, for the target use case, be added to the inner netlist.
The default value for both parameters is false.  These parameters are
only recognized from the "second level manager" class, which means that
they would be ignored by conventional simulations, that do not use this
algorithm.  They also will not be recognized by the top level solver, so
if the user inadvertantly puts them in the top-level netlist, they won't
do anything.

For the target use case, the vector output (F,Q,B) is the main
interest.  The matrix output, while interesting, may not always be
desired.  This output capability can lead to a really large number of
files (a separate set at every time step), so best to reduce it when
possible.  So, I expect that in most cases, the only command needed
will be

.options twolevel output_dae_vectors=true

In addition to having the option to only output vectors, these
new functions have the following advantages over the old debug
functions:

1.  They are not debug functions, so they can be invoked from any build
    of Xyce.  It is not necessary to build a debug version.

2.  They are only called on the inner solver at the end of a successful
    outter solve.  So, they will only output a single set of DAE vectors
    and/or matrices for each time step.