Remove block gainscale "homotoy" and randomizevgst
In commit Xyce@2e8028c6 back in 2005 Roger P. added a new continuation
algorithm that randomly placed devices that support it into different
blocks, and then smoothly modified the MOSFET "gainscale" parameter
for each block from 0 to 1.

This commit removes that feature and an associated hack.

According to the commit message:

"Adding a few new homotopy methods to xyce.  They don't work too well,
but I'm adding it so people canplay with it and the infrastructure
could be used in the future for more types of continuation."

This method that doesn't work so well was put into Xyce in 2005, and
has not been used since.  The infrastructure it implemented has not
been used for more types of continuation, either.

The use of random numbers in the original implementation introduced a
dependence on a Trilinos header file in one of the Xyce header files
needed by codes that couple to Xyce, making it difficult to distribute
precompiled binaries of Xyce without including a full Trilinos build
with it.  I had to rework how random numbers were obtained without
resorting to Trilinos, in commit fe83aae8.  That led to a discussion
about how this "block gainscale homotopy" was not in fact ever used
and could be removed altogether.

This was not as simple as just reverting commit Xyce@2e8028c6, because
17 years of code development were in the way.  Furthermore, an
additional hack was put into place shortly afterwards that allowed
further randomization of MOSFET parameters during the artificial
parameter continuation ("RANDOMIZEVGST") in commit  1bd8850f. That
commit's primary purpose was to add a sort of pseudo-transient
continuation, but threw RANDOMIZEVGST in for free.  I removed
RANDOMIZEVGST, too.

There was a slight obfuscation introduced later when some analysis
package code that involved this feature was refactored, but the end
result was that that refactor made it *easier* to excise block
gainscale homotopy than it would have been before the refactor.  Whew.

No tests were ever created for this unused feature, and it was never
documented in the users or reference guides.  So nobody was ever told
what continuation type "7" did, and nobody needs to be told it doesn't
exist anymore.  There will be no user or reference guide updates, nor
will this be mentioned in release notes.

The main reason for wanting to get rid of this feature is to get rid
of one more use of the RandomNumbers class that I had put into Xyce
when we couldn't use C++11 random number functions and needed more
than just the system provided generator.  Now that we require a C++11
compiler, there is no longer a good reason to keep this class around
other than that we still use it in a few places.  One fewer, now.

Xyce#57
Xyce#58
Xyce#56