Remove RandomNumbers usage from MemristorTEAM
The MemristorTEAM device has an option that enables random numbers to
modify the resistance ("resistor noise").  It uses uniformly and
Gaussian distributed numbers while setting this up.

This commit switches its use of the RandomNumbers class, which was
always sketchy since it uses the system-provided random number
generator, which has a single seed.  Any use of this class that
reseeds the generator *also* causes all other users of the class to
have their random number sequence changed out from under them.

We went through some lengths to centralize initialization  of the seed
consistently across uses of RandomNumbers (so that the "-seed" command
line option would actually work), but repeated use of the class all
over the code (with duplicated calls to seed using different methods)
breaks that.

Anyhow, we want to get rid of the RandomNumbers class.  This commit
gets rid of it in this one device.

Unfortunately, there is not a single test in the test suite that uses
the option that enables random noise in this device.  So it is
impossible to verify either that it was ever correct, or that it is
now correct after modification.  However, it does use the C++11
standard random functions instead of homegrown stuff.  And the seed
used in this device is truly local to this device.

I can see already that there is a broken behavior of this device if
the random feature is enabled and no seed is given.  In that case it
simply uses 0 as the seed.

What *should* happen, and what happens in all other uses of random
numbers in the code, is that if no seed is given then a suitable
random seed is generated.  In most cases before we used "time(NULL)"
for the seed.  Now, we should use "std::random_device rd;" and pass
"rd()" to the "mt19937" object as the seed when it is constructed.

I have NOT fixed that, because it was never right before and there are
no tests.  But I have added comments in the appropriate place, and
when the time comes to test this aspect of this device, it can be
fixed then.

Xyce#58