Fix errors with AMD CMakeLists.txt script
The CMakeLists.txt AMD build script is supplied as a convenience for building
AMD from a full SuiteSparse repo. It builds only AMD, so it is faster than
building the entirety of SuiteSparse, and is designed to integrate with a
Trilinos build.

When initially committed, the name of the script had a lower case 'm' as the
second letter, so CMake would fail to find it on a case-sensitive OS. It also
did not specify the type of build, so I set CMAKE_BUILD_TYPE to "Release".
Finally, I took the opportunity to improve some of the comments in the script
describing how it works and how to use it.