Remove RandomNumbers usage from Synapse3 device
We're removing the RandomNumbers utility class and replacing use of it
with C++11 standard random number routines.

This commit removes this usage from the Synapse3 device in a manner
that is consistent with its former operation.  Specifically, there is
a single random number generator as a static class member, and it is
seeded once.  All instances pull their random numbers from this same
pool.

I'm not sure how well that technique plays in parallel operation, but
the new code will be no worse than the old in that regard.

There is one test of the synapse3 device in the regression suite, and
it exercises the random number usage (according to our coverage
data).  It is passing after this change.

Xyce#58