Isolate Matlab search behind a cache variable
For the Simulink interface, CMake unconditionally searches for Matlab
and enables the interface if found. The search is not robust enough for
reular use, though, so I've put it behind a cache variable,
'Xyce_SIMULINK'. To enable the search, the variable has to be turned on
at the CMake invocation.