Fix broken noise contributions with analog function calls
During his investigation of #392, @erkeite found that some Verilog-A
models he obtained from github would not compile in Xyce/ADMS without
a workaround.  I tracked this down to improper handling of noise
contributions whose RHS contained analog function calls.

This commit fixes such noise contributions.  The issue was that in the
early days of our desacadofication of Xyce/ADMS, I had made a decision
regarding noise contributions that was not strictly necessary, nor was
it completely debugged.  Undoing that decision was easier than
debugging it.

The main issue is that I had decided not to do function precomputation
for noise contributions, and had to signal to the code generator not
to assume that function precomputation was being done.  However there
was a bug in there so that it was still assuming function
precomputation was being done for analog functions.  Now it just does
precomputation.

Xyce#396