Sanity check xyceModelGroup attribute
Xyce/ADMS allows a user to specify attributes on Verilog-A module
definitions that amount to directives about how Xyce should plug the
device into its model library.  For example, setting the
'xyceModelGroup' attribute to 'BJT' causes the device to be usable as
a "Q" device in Xyce instead of as a Y device.

However, the code that handled this in the Xyce/ADMS back-end did no
sanity checking on this, and if a user specified a model group that
doesn't even exist in Xyce, then nonsensical code would be emitted.

Now Xyce/ADMS will exit with an informative error message telling the
user that what they've tried is wrong, and how to fix it.

Xyce#399