Fixes CMake test and ifdef for ShyLU-Basker solver
The easiest way to test for the ShyLU-Basker solver is through the
Amesos2 interface, instead of through ShyLU.  This is mostly because
we use the ShyLU-Basker solver through the Amesos2 interface, so it
is more important that Amesos2 has the solver enabled than ShyLU.

The check enables a new Xyce_AMESOS2_SHYLUBASKER definition in the
Xyce_config.h.cmake file that is used in the Amesos2 interface to
make the ShyLU-Basker solver accessible.  There was also some cleanup
needed in the Amesos2 interface to remove debug statements.

With these changes, Xyce will detect and correctly build the binary
to use the ShyLU-Basker solver with CMake.  This has been tested on
the development version of Trilinos, which has been hardened for
usage in Xyce.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/234