Cleanup 2-level DAE vector outputs
Xyce/Xyce#379

This slightly cleans up the new DAE outputs, which are used with 2-level
solves.  It also corrects an indexing mistake, that happened when used
with a non-debug build of Xyce.  This indexing mistake affected the file
names.

Now, with this change, non-debug and debug builds will produce the same
list of files when:

.options twolevel
+ output_dae_vectors=true
+ output_dae_matrices=true

is present in the netlist.