Address coding concerns brought up in QA
In issue #58, Josh brings up some coding style questions.

I have violated a long standing convention in Xyce, that standard
headers be included *first*, and another that standard headers be
included in the file that requires them and not depend on them showing
up as a consequence of some other Xyce or Trilinos include.

This commit fixes those convention violations.

Xyce#58