More cleanup 2-level DAE vector outputs
Xyce/Xyce#379

This commit changes the output file names, so that only the time step
index is used as part of the name.    Preivously, several other indices,
including (problematically) the Newton step number, were included in the
name.  As these new functions are ONLY outputting at the end of a
converged Newton loop, the step number should not be part of the name.

The other two indices (continuation step number and parameter step
number) have been removed as well.  For the use case of interest they
were always zero.