Add warning to AC source for ACMAG=0
I was recently given a circuit to debug, and one issue that it had was
that the AC source magnitude was set to zero.  Xyce didn't have an
error trap for this issue, and computed an answer.  But with the only AC
source set to zero, that meant the AC right-hand-side vector was
all zeros.

Unlike a DC source, where there is a known use case for a zero-valued
source, I can't think of a valid use case for a zero-valued AC source.
So, I have added a warning to the source class, that is emitted when an
AC source is set to zero.

There is no issue associated with this commit.