Fix for issue #405
This commit fixes Issue #405 :
"Fix inline subcircuit param dependencies"

This fix is also relevant to issue #267, which is about the M=
multiplier for subcircuits.  This commit does not fix that issue, but
is in the same area of the code.

There are 3 types of "subcircuit parameters" that are affected by this
fix/refactor.  They are:

1. .param inside the subcircuit
2. parameter on the subcircuit definition line (.subckt PARAMS:)
3. parameter on the subcircuit instance line ("X" lines)

For the above types, (1) and (2) are basically the same thing and should
be treated 100% the same.  In the old code they were not, and this led
to various corner use cases not working correctly.

In this commit, (1) and (2) are now treated the same, with ".subckt PARAMS:"
being the equivalent of a .param.  For the purposes of precedence, if a PARAMS:
parameter matches a .param inside the subcircuit, the PARAMS: parameter is
considered to have appeared "first" in the scope.

Which is used (the first or last) depends on the the precedence rule
command that I implemented for issue #202.   That command is:
Xyce -redefined_params <option> where <option> can be
IGNORE, ERROR, WARNING, USEFIRST or USEFIRSTWARN.  IGNORE (or USELAST)
is the default behavior.

Parameter type (3) is fundamentally different than (1) and (2).  It is
different in that it is in the parent subckt scope, rather than the
child scope of (1) and (2).  Also, if it duplicates a (2) or (3)
parameter, it ALWAYS takes precedence, no matter what setting is used
for -redefined_params.

So, the fix for this was to:

1. create a new function, DeviceBlock::setSubcircuitInstanceParameterValues().
   Previously, device parameters (including subcircuit X parameters) were
   always resolved by DeviceBlock::setParameterValues().  This way,
   subcircuit X parameters are allowed to refer to each other. (this is
   not true for non-subckt devices).

2. delete most of the CircuitContext::resolve code handling .subckt PARAMS:
3. delete most of the CircuitContext::resolve code handling "X" parameters.

4. put .subckt PARAMS: into the unresolvedParams_ object, and resolve the
   same as any other .PARAM.  Apply the -redefined_params precedence
   rule.

5. Put the "X" params into the "resolvedParams" object.  If they match
   any params in the "unresolvedParams" object, delete them.

6  After this, resolve all parameters in the unresolvedParams object
   like before, using a while loop that keeps calling the
    "CircuitContext::resolve" function.

Now a line like this works:

Xsub node1 node2 test PARAMS: par1=2.0 par2='3*par1'

whereas it did not work before.

Also lines like this:

.global_param gp1=5.0
.subckt simple in out PARAMS: par1=gp1  par2=2.0 par3='par1*par2*2.0'

work more correctly than before.  In the old code this worked fine if
gp1 was constant, but it didn't update par3 correctly if gp1 was swept
using .STEP. Now it does.