Add param size check for AC SENS
When setting up to do sensitivity analysis, the AC class does an error
check to make sure the option block includes at least one objective
function, but there is no check on the number of parameters for the
objective(s). If no objective function parameters are defined, running
the sensitivity analysis will cause Xyce to crash. The commit adds a
check for objective function parameters.

Add objective function check for DC/TRAN SENS

When setting up to do sensitivity analysis, the DC/Tran class does an
error check to make sure the option block includes at least one
function parameter, but there is no check on the number of objective
functions for the analysis. If no objective function(s) are defined,
Xyce still runs through the simulation, but it doesn't produce any error
messages or sensitivity output. The commit adds a check for missing
objective function parameters and treats it like an error.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/371