Begin process of revamping release creation
Our release process has involved running a script called
"make_release" ever since Xyce 1.0.  This script has always been a
kludge, and was originally created when we had to bundle an enormous
shared library from Comet Solutions, and when we  used a "runxyce"
script to hackishly make sure the Xyce binary could find it (instead
of properly using RPATH in the binary).  This script cherry picked
individual files out of the build directory and bundled various Intel
compiler libraries with it if necessary.  It was a bit of a mess, and
still had remnants of when we were supporting DEC Alpha, Solaris, and
SGI.  It also had some kludges regarding how we would bundle
documentation with the binaries, which we stopped doing several
releases ago.

We've been stuck with that for 20 years.

I am trying to revamp the process to use "make install" to create a
complete install that includes Xyce libraries and headers and which
could, in theory, support our distributing shared library builds.

It's complicated for a number of reasons, but the script itself is
really simple if you throw away a lot of the old kludges.

This commit is the first step.  It creates a new
"make_release_tarball_installer" that creates a simple installer
similar to the one we have always created, but instead of cherry
picking individual files out of the build directory bundles everything
that "make install" would install.  It doesn't do that make
install *or* configure Xyce with an appropriate prefix, but rather
assumes that it's already been done --- you just tell it where it was
installed.

Xyce#406