Add note about packaging change to release notes
Though largely of interest to nobody (because it doesn't yet provide any
immediate improvement in function), Xyce binary packages for OS X and
RHEL7 now contain lots more files than before, and a user with just
the right environment (the same compilers we have, and a version of
Trilinos built exactly how we build it for releases) could use the
libraries and headers contained here to link to Xyce.

They need the same compiler because the one bundled with RHEL7 won't
work, and they need Trilinos because that's how static libraries
work.  Were they shared libraries they'd need only the compiler to be
right, and Trilinos dependencies would be resolved from libxyce.so at
run time instead of at link time.

Anyhow, users need to know what changed, and this commit tells them in
the release notes.

Xyce#406