Bundle more MKL libraries
During QA, Rich discovered that on his system Xyce required an MKL
library other than the one bundled with the code.  This was
libmkl_intel_lp64.1.dylib.  Only libmkl_intel_lp64.dylib was bundled.

There are two other mkl libraries bundled, and both of these have ".1"
variants in the MKL library directory.

This commit causes all of the variants to be copied, not just the one.

This could not be caught in normal testing because we test the code on the
same machine where it is built, which has all the libraries.  When our
system was upgraded to use the OneAPI compiler it may be that this new
dependency was pulled in, and our binary installer may never have been tested
properly after that.

Xyce#406