Set RPATH unconditionally
Years ago, I added code to configure.ac to add an RPATH setting only
when we absolutely knew we needed one, which was when using Intel
compilers.  There was conditional code to figure out if we were using
the Intel compiler based on the setting of CXX, and if so set the RPATH.

This is no longer correct.  We now build Xyce on OS X with Clang, but
using the Intel MKL libraries --- so we still need to bundle the
libraries, and absolutely do still need to set an RPATH.  We were not.

This is a new change since Xyce 7.4, and happened when we upgraded our
testing Mac to the latest OS --- that broke our old Intel
compiler *and* our old packaging system.  @dcollin changed the build
to use clang, but clearly we never tested that the packaged binary
worked.  This bit us during QA of Xyce#406.

Setting an RPATH doesn't hurt even if we don't need one, so I am
changing the configure script to do so no matter what.  For Linux and
BSD, the RPATH is always added to one that is relative to the Xyce
binary (using $ORIGIN), but OS X does not have an equivalent of that.
So I'm hard coding an RPATH of /usr/local/${CPACK_PACKAGE_NAME}/lib.

Libtool will set an rpath automatically if we are building a
SHARED build, but we're not doing that here.

Xyce#406